platform: msm: i2c_qup: fix multipile blsp usage during execution
the i2c_blsp qup module is using a global to handle interrupts, it is
set during init function, while the next init will fail in case the
global is not null.
the deinit fanction lacked the reset of the global, in order to allow
the next i2c_blsp init call.

Change-Id: I0cd6e0fb150fffbc52829d0f7414be0ac2dea2e0
Signed-off-by: Ido Reis <ido.reis@tandemg.com>